src/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
google-auth
google-cloud-storage
